Upgrade enterprise generation with pipeline only support
Previously it was not possible to generate a page
without data from a gitlab server. With this change,
pipeline only is supported. The implemented mechanism
that leads to the generation of a enterprise page is currently
still very inefficient. MongoDB is not optimally implemented yet.

Also there were a structural changes which includes moving the
page generation algorithms from models.py to services.py